[chore] 충돌 발생한 클래스 확인 및 해결
